Add URL shortener proxy API endpoint
The final decoded URLs are too long for Denon AVR devices using the HEOS
command line interface. The underlying CLI system has a maximum of 512
bytes for the URL passed in to the `heos://browse/play_stream` command
and anything over that will crash the HEOS service abruptly (is there
any other kind) and require a reconnection to send commands.

This adds an optional advanced configuration to enable an API endpoint
for Home Assistant to act as a simple redirect proxy by creating a much
shorter URL that will issue a 302 redirect to the real service. It turns
out the underlying player _can_ handle a longer URL but not via the HEOS
command.

The player adds a redirect dictionary to allow for a mapping of the MD5
of the final URL to the URL itself and replaces the URL with
`/api/ytmp/s/{player-id}/{md5}` and sends that to the player instead.
The API endpoint then extracts the player object and finds the
appropriate URL to send as a redirect. When the YouTube Music player is
turned off the cache is cleared, being an easy compromise to keep the
cache under control without adding overheads of tracking the age of the
URL and periodically cleaning them out. Of course, if you never turn off
the player it will grow unbounded until Home Assistant is restarted.

An earlier implementation of this created a new `aiohttp` server on
another port, which worked for me locally but may have had issues
running in the full docker environment and by enabling the API endpoint
it technically works with the cloud service and any HTTPS endpoints you
may have configured. I have no idea how it would handle a self-signed
cerificate though, so that may still be an issue.