Only enable the HTTP proxy if requested
A small change to only register our HTTP view if a player has been
configured with the option enabled. This uncovered some issues if the
setting was changed via the options form so that was cleaned up as well.
This change also moved the `short_proxy_url` string to a constant so
it's more consistent.